Fixed case testing
most of the run.py files for the tutorials encountered a TabError following the conversion to Python3. This fixes this error (mostly changes TAB+space to just TAB), as well as other minor errors with unit testing regarding type matching for np.linspace.

+++ bemio/io/nemoh.py
@@ -168,7 +168,7 @@ class NemohOutput(object):
 
         # Read wave directions
         temp = cal[-6]
-        self.cal.wave_dir_n = np.float(temp.split()[0])
+        self.cal.wave_dir_n = int(temp.split()[0])
         self.cal.wave_dir_start = np.float(temp.split()[1])
         self.cal.wave_dir_end = np.float(temp.split()[2])
         self.cal.wave_dir = np.linspace(self.cal.wave_dir_start,self.cal.wave_dir_end,self.cal.wave_dir_n)
